ScoreDetailContent を分離
